testing/web-platform/tests/XMLHttpRequest/allow-lists-starting-with-comma.htm

<!DOCTYPE html>
<html>
  <head>
    <title>Allow lists starting with a comma should be parsed correctly</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
  <script type="text/javascript">
    var test = async_test();

    test.step(function() {
      var xhr = new XMLHttpRequest();
      var url = "resources/access-control-allow-lists.py?headers=,y-lol,x-print,%20,,,y-print&origin=http://127.0.0.1:8000";
      xhr.open("GET", url, false);
      xhr.setRequestHeader('x-print', 'unicorn')
      xhr.setRequestHeader('y-print', 'narwhal')
      // Sending GET request with custom headers
      assert_equals(xhr.send(null), undefined);
      var response = JSON.parse(xhr.response);
      assert_equals(response['x-print'], "unicorn");
      assert_equals(response['y-print'], "narwhal");

      url = "resources/access-control-allow-lists.py?methods=,,PUT,GET&origin=http://127.0.0.1:8000";
      xhr.open("PUT", url, false);
      // Sending PUT request
      assert_equals(xhr.send(null), undefined);
      test.done();
    });
  </script>
  </body>
  </html>
